Move tests
They exercise the new validation infrastructure, so it makes sense to
move them there.
